Fix min/max String error bug
When you provide an image of a type that can be negative, the Op can
attempt to access the string of valid ascii image characters at a
negative index, throwing an error. This bug was fixed and a test was
added with the code that threw an error before, as well as making sure
that the clamped ascii image is not the same as the unclamped image.