DefaultASCII: reorganize math
Note: if the RealType of the image is not a type supporting floating point
numbers, then div(span) on that type would cause a rounding error. This
is the reason that normalized has to call getRealDouble() twice,
although it seems it would be nicer if we used the Imglib2 Type Math.