Add a seam to replace hardcoded threads to use
We add a getMaxThreads() method to OpEnvironment, and use it everywhere
Runtime.getRuntime().availableProcessors() was previously being used.

The default is still the available processors (which according to the
spec can actually vary over time within the same runtime environment,
although I'm not sure if it ever actually does in any real JVM
implementation). But now it can be overridden by calling
setMaxThreads(int) to make it fixed.